+++ src/azure-firewall/HISTORY.rst
@@ -3,6 +3,10 @@
 Release History
 ===============
 
+2.1.1
+++++++
+* Update AzureFirewall model to support extended location.
+
 2.1.0
 ++++++
 * `az network firewall policy intrusion-detection add`: Update IDPS profile handling to use the latest supported profile values (`Off`, `Emerging`, `Core`, `Extended`).